chore: extract dockerignore to the image package
