+++ image/testdata/.dockerignore
@@ -1,0 +1,2 @@
+Dockerfile
+echo.Dockerfile
