+++ image/testdata/dockerignore/.dockerignore
@@ -1,0 +1,5 @@
+# .dockerignore
+# https://docs.docker.com/build/building/context/#dockerignore-files
+vendor
+foo
+bar
